Rollback SymPy version to fix serious errors occurring on first parse
A good example of the weirdness in SymPy 1.5 and above can be seen trying:
```
maths.check("sqrt(1 - 1/sqrt(2))/sqrt(2)", "sqrt((2 - sqrt(2))/4)")
```
which leads to
```
CoercionFailed: expected `Rational` object, got 1 - sqrt(2)/2
```
the first time you attempt to run it, but subsequent attempts work fine.
This is not the only expression, unfortunately, but I didn't write down any
others. It's almost like a bad cache or badly written try except block?
At any rate, the issue is not seen in SymPy 1.4