Fix precedence issue with boolean keywords and equals
Boolean keywords bind less tightly than ==, to allow for coding things
like "var_a == var_b and var_c" to be meaningful. But we need Python to
parse "A == B or C" as an equation, so equals must not be tighter.
The only way to achieve this easily is to transform "and", "or" etc
during tokenisation into the bitwise operators, which do bind tighter
than equals. We can then transform them back into boolean things as we
already do during the AST rewriting.