Add 8000+ real-world test cases via PyTest parameterisation
We clearly don't want to run these every time, so they do not live in
the module itself, but they are very useful for more thoroughly
validating the behaviour of the checker.
They are all genuine expressions submitted by learners as equivalent to
the correct expression, as decided by the v0.12.0 release.

Run PyTest in the top-level directory to run both the unit tests and
this test suite, or just run this file itself with:
  pytest test_cases.py